requirements.txt
SpeechRecognition==3.8.1
pyttsx3==2.90
anthropic==0.3.6
opencv-python==4.10.0.84
numpy==2.1.2
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
